+++ inst/scripts/toy_example_examples.r
@@ -1,0 +1,27 @@
+library(HMVAR)
+library(tidyverse)
+
+# Get file paths
+midas_dir <- system.file("toy_example/merged.snps/", package = "HMVAR")
+map <- read_tsv(system.file("toy_example/map.txt", package = "HMVAR"),
+                col_types = cols(.default = col_character())) %>%
+  select(sample = ID, Group)
+
+# Read data
+midas_data <- read_midas_data(midas_dir = midas_dir, map = map, cds_only = TRUE, genes = NULL)
+midas_data
+
+########################
+
+library(HMVAR)
+library(tidyverse)
+
+# Get file paths
+midas_dir <- system.file("toy_example/merged.snps/", package = "HMVAR")
+map <- read_tsv(system.file("toy_example/map.txt", package = "HMVAR"),
+                col_types = cols(.default = col_character())) %>%
+  select(sample = ID, Group)
+
+# mktest
+midas_mktest(midas_dir = midas_dir,
+             map_file = system.file("toy_example/map.txt", package = "HMVAR"))